Move home constants to global constants folder & refactor root animation duration to theme layer
