+++ index.html
@@ -41,14 +41,14 @@
     <title>Quan Tran</title>
   </head>
   <body>
-    <div id="splash-screen" class="flex-center splash-screen">
+    <!-- <div id="splash-screen" class="flex-center splash-screen">
       <div class="flex-col-center splash-content animate-fade-in" style="animation-duration: var(--fade-duration);">        
           <img src="/splash-logo.png" alt="splash-logo" class="splash-logo" />
           <div id="loading-bar-container" class="progress-bar-container">
             <div id="loading-bar" class="progress-bar animate-loading-bar" style="animation-duration: var(--loading-bar-animation-duration);"></div>
           </div>
       </div>
-    </div>
+    </div> -->
 
     <div id="modal"></div>
     <div id="root"></div>
