+++ src/components/ui/RouteLoadingIndicator.tsx
@@ -1,22 +1,17 @@
-import { RefObject } from "react";
-
 type RouteLoadingIndicatorProps = {
   className?: string;
   stagger: number;
   dotCount: number;
-  ref?: RefObject<HTMLDivElement | null>;
 };
 
 export default function RouteLoadingIndicator({
   stagger,
   dotCount,
   className,
-}: // ref,
-RouteLoadingIndicatorProps) {
+}: RouteLoadingIndicatorProps) {
   return (
     <div
       id="loading-indicator-screen"
-      // ref={ref}
       className={`size-full z-loading bg-light absolute inset-0 ${className}`}
     >
       <div className="h-full flex-col-center gap-[clamp(0.75rem,2vw,3rem)]">
@@ -25,7 +20,7 @@ RouteLoadingIndicatorProps) {
           {Array.from({ length: dotCount }).map((_, index) => (
             <div
               key={`${index + stagger}-loading-dot`}
-              className="bg-dark rounded-full animate-bounce-fade size-[clamp(0.75rem,1.25vw,2rem)]"
+              className="bg-dark rounded-full animate-bounce-fade  size-[clamp(0.75rem,1.25vw,2rem)]"
               style={{
                 animationDuration: `var(--bounce-fade-animation-duration)`,
                 animationDelay: `${index * stagger}ms`,
