+++ src/layouts/RootLayout.tsx
@@ -1,5 +1,4 @@
 import { Outlet, useNavigation } from "react-router-dom";
-import { useRouteLoadingIndicator } from "@/hooks/useRouteLoadingIndicator";
 import { getRootStyleProperty } from "@/utils/dom-utils";
 import RouteLoadingIndicator from "@/components/ui/RouteLoadingIndicator";
 import Footer from "@/components/layout/Footer";
@@ -9,8 +8,7 @@ import { toMilliseconds } from "@/utils/time-utils";
 import { useRemoveSplashScreen } from "@/hooks/useRemoveSplashScreen";
 
 export default function RootLayout() {
-  useRemoveSplashScreen();
-  // const { ref } = useRouteLoadingIndicator();
+  // useRemoveSplashScreen();
   const loaderDot = useMemo(
     () => ({
       count: getRootStyleProperty("--loader-dot-count"),
@@ -27,7 +25,6 @@ export default function RootLayout() {
       <main className="relative flex-center grow px-[clamp(1.25rem,calc(12.5vw_-_2rem),20rem)] py-[clamp(1rem,1vw,3rem)]">
         {isNavigating && (
           <RouteLoadingIndicator
-            // ref={ref}
             dotCount={parseInt(loaderDot.count)}
             stagger={toMilliseconds(loaderDot.dotStagger)}
             className=""
