+++ src/pages/home-page/index.tsx
@@ -1,8 +1,8 @@
 import HeroBanner from "@/pages/home-page/components/HeroBanner";
 import HeroContent from "@/pages/home-page/components/HeroContent";
-import { HOME } from "@/pages/home-page/home.constants";
+import { HOME } from "@/constants/home.constants";
 
-export function Home() {
+export default function Home() {
   return (
     <section id="home">
       <div className="grid gap-[clamp(3.75rem,4.5vw,6.25rem)] lg:grid-cols-2">
